[Java.Interop.Tools.JavaCallableWrappers] Create common method for validating identifier names.
