+++ src/Java.Interop.Tools.JavaCallableWrappers/Java.Interop.Tools.JavaCallableWrappers.csproj
@@ -34,6 +34,7 @@
     <Reference Include="System" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Java.Interop.Tools.JavaCallableWrappers\IdentifierValidator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Java.Interop.Tools.JavaCallableWrappers\Crc64.cs" />
     <Compile Include="Java.Interop.Tools.JavaCallableWrappers\JavaCallableWrapperGenerator.cs" />
