+++ src/Java.Interop.Tools.JavaCallableWrappers/Test/Java.Interop.Tools.JavaCallableWrappers-Tests.csproj
@@ -39,6 +39,7 @@
   <ItemGroup>
     <Compile Include="Java.Interop.Tools.JavaCallableWrappers\JavaCallableWrapperGeneratorTests.cs" />
     <Compile Include="Java.Interop.Tools.JavaCallableWrappers\Crc64Tests.cs" />
+    <Compile Include="Java.Interop.Tools.JavaCallableWrappers\IdentifierValidatorTests.cs" />
     <Compile Include="Java.Interop.Tools.JavaCallableWrappers\JavaNativeTypeManagerTests.cs" />
     <Compile Include="Java.Interop.Tools.JavaCallableWrappers\TypeNameMapGeneratorTests.cs" />
     <Compile Include="Java.Interop.Tools.JavaCallableWrappers\SupportDeclarations.cs" />
