+++ tools/generator/Utilities/TypeNameUtilities.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace MonoDroid.Generation
@@ -102,41 +101,5 @@ namespace MonoDroid.Generation
 			}
 			return builder.ToString ();
 		}
-
-		// From https://github.com/xamarin/xamarin-android/blob/master/src/Xamarin.Android.Build.Tasks/Utilities/ResourceIdentifier.cs
-		// https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/language-specification/lexical-structure#identifiers
-		private const string FormattingCharacter = @"\p{Cf}";
-		private const string ConnectingCharacter = @"\p{Pc}";
-		private const string DecimalDigitCharacter = @"\p{Nd}";
-		private const string CombiningCharacter = @"\p{Mn}\p{Mc}";
-		private const string LetterCharacter = @"\p{Lu}\p{Ll}\p{Lt}\p{Lm}\p{Lo}\p{Nl}";
-
-		private const string IdentifierPartCharacter = LetterCharacter +
-		    DecimalDigitCharacter +
-		    ConnectingCharacter +
-		    CombiningCharacter +
-		    FormattingCharacter;
-
-		private const string IdentifierStartCharacter = "(" + LetterCharacter + "_)";
-
-		private const string Identifier = IdentifierStartCharacter + "(" + IdentifierPartCharacter + ")";
-
-		// We use [^ ...] to detect any character that is NOT a match.
-		static Regex validIdentifier = new Regex ($"[^{Identifier}]", RegexOptions.Compiled);
-
-		public static string CreateValidIdentifier (string identifier, bool useEncodedReplacements = false)
-		{
-			if (string.IsNullOrWhiteSpace (identifier)) return string.Empty;
-
-			var normalizedIdentifier = identifier.Normalize ();
-
-			if (useEncodedReplacements)
-				return validIdentifier.Replace (normalizedIdentifier, new MatchEvaluator (EncodeReplacement));
-
-			return validIdentifier.Replace (normalizedIdentifier, "_");
-		}
-
-		// Makes uglier but unique identifiers by encoding each invalid character with its character value
-		static string EncodeReplacement (Match match) => $"_x{(ushort) match.Value [0]}_";
 	}
 }
